Main function created, added Cohens D score, PCA and logger
Wrote a new script which produces a pandas data frame and log file of statistical tests and also provides information on the data.
Added a logger, some checks of data and group size, Cohens ds score and batch correlation with the first 3 PCAs.
Key results are then stored in the pandas data frame and summaries printed in the log_file