Added more diagnostic functions and plots
Added Mahalanobis distance with pairwise and centroid distance, including with covariates in the model and after residualising as well as plotting function.
Added variance ratio + plotting function.
Deleted old Logger tool and replaced with new version that writes output as an HTML file, allowing figures to be saved to the file.
Began putting the pipeline for cross-sectional data (DiagnosticReport.py) together.

Added test functions for the new additions and refined old test functions.