Big push of the harmonization tool pipeline
Added a working version of the harmonization tool that combines each of the tests into one function and returns a detailed HTML file with plots embedded into it.

The report takes the following arguments:
        DiagnosticReport.DiagnosticReport(
            data=data,
            batch=batch,
            covariates=covariates,
            covariate_names=variable_names,
            save_dir=str(out_dir),
            SaveArtifacts=True,
            rep=None,
            show=False
        )

And returns each of the plots as a PNG and a HTML report with each of the major plots and a logging for the tests. This part needs work as currently it only writes the statistical test results for mahalanobis (batch vs batch and batch vs centroid) and Cohen's D (batch vs batch). Cohen's D plot needs to be refined such that it returns the score vs each batch instead of just the last one in future works. (i.e gives the number of features that are significant per batch rather than just total number of features)

Added in a logging tool which manages this and creates the HTML, design needs to be changed an additional logging.break added to make the report look pretty.

Asked ChatGPT 5 to make an inbuilt wrapper to work with the logging tool to allow the plots to be added to the report without changing any of the original plotting functions.

Bugs: 'show=False' flag not working for KS_test plot as of yet, unsure of why this is as the wrapper works well for PCA_corr
Cohens D score is summary instead of batch wise issue, needs to be refined so that we have single Cohens D plot for each batch vs global mean as with Mahalanobis.

Other changes:

Renamed testing functions to be more consistent
Changed some test functions and added one for the Diagnostic report
Changed some of the diagnostic function so they can take lists of strings rather than numeric or categorical vectors for batch (simple fix of just creating a new array from the list that is numerical, only for answering questions in PCA or LMEM where the specific batch difference isn't important but the variance explained by the batch array is)